Ensure that the action bar does not overlap the fragments
To ensure that fragment content does not disappear under the action bar, we add a marginBottom with exactly the height of the actionBar.
Please note that it is pretty unexpected that this measure is necessary, as the <fragment> is configured withn a constraint to be on top of the actionBar "@id/nav_view" in activity_parkview.xml:
   app:layout_constraintBottom_toTopOf="@id/nav_view"
Alas, this configuration does not work, event though AndroidStudio auto-generated that layout. Several people report issues with thsi, especially if a ScrollView is part of the fragment. THe solution that worked is from https://stackoverflow.com/questions/36810873/navbar-overlapping-last-listview-item-android. This may or not be the perfect solution, as there may be issues on "devices without soft navbar". See https://stackoverflow.com/questions/27723322/bottom-soft-navigationbar-overlaps-my-listview for more potentially working solutions.